+++ src/main/java/Point.java
@@ -1,8 +1,8 @@
 public class Point {
-    public final double x;
-    public final double y;
+    public final float x;
+    public final float y;
 
-    public Point(double x, double y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
